Insert ethernet header for packets from tunnel interface and destined to vhost
As part of receiving packets on the tunnel interface, make sure that there
is enough head space in the packet to add ethernet header + agent headers.
All packets ingressing the vhost interface needs an ethernet header. For
packet arriving on a tunnel interface and destined to vhost, ethernet header
will not be present, and hence need to be inserted